pyproject.toml
[project]
name = "life-timeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=23.1.0",
    "flask>=3.1.0",
    "flask-admin>=1.6.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=23.0.0",
    "passlib>=1.7.4",
    "python-dotenv>=1.0.1",
    "python-slugify>=8.0.4",
    "wtforms>=3.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.0",
]
